-CActionRegTicket::GetActionFees returns map of action ticket fees: calculated sense and cascade fee
  mnode-controller:
   - moved TRIMMEAN implementation to the separate file trimmean.cpp
   - added default action ticket fee (10) and GetActionTicketFeePerMB
   - refactored GetNetworkFeePerMB / GetNFTTicketFeePerKB / GetChainDeflationRate
   - TRIMMEAN refactored: removed memory allocation for array, replaced with data vector, fixed data types for indexes
  mnode-rpc:
   - moved and refactored storagefee rpc command to the separate file mnode/rpc/storage-fee.cpp|handled
   - added implementation for "storagefee getactionfees <data_size>"
   It returns json array of action fees:
[
    { "action_type": "<action_type_name>",   (string) action type name
      "fee": x.xxx,                          (numeric) fee in PSL
      "feePat": xxxx                         (numeric) fee in patoshis
    },
    ......
]

  google tests:
   - added tests for TRIMMEAN implementation

+++ src/mnode/rpc/masternode.cpp
@@ -279,7 +279,7 @@ UniValue masternode(const UniValue& params, bool fHelp)
         std::string strMode = params[1].get_str();
 
         if (strMode == "enabled")
-            return masterNodeCtrl.masternodeManager.CountEnabled();
+            return static_cast<uint64_t>(masterNodeCtrl.masternodeManager.CountEnabled());
 
         int nCount;
         masternode_info_t mnInfo;
@@ -289,7 +289,7 @@ UniValue masternode(const UniValue& params, bool fHelp)
             return nCount;
 
         if (strMode == "all")
-            return strprintf("Total: %d (Enabled: %d / Qualify: %d)",
+            return strprintf("Total: %zu (Enabled: %zu / Qualify: %d)",
                              masterNodeCtrl.masternodeManager.size(), masterNodeCtrl.masternodeManager.CountEnabled(), nCount);
     }
